Rename directory: gin-client -> ginclient
Package name was already ginclient. Changing path to match.
