+++ gincmd/annexcmd.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"os"
 
-	ginclient "github.com/G-Node/gin-cli/gin-client"
+	"github.com/G-Node/gin-cli/ginclient"
 	"github.com/G-Node/gin-cli/util"
 	"github.com/spf13/cobra"
 )
