+++ gincmd/infocmd.go
@@ -4,7 +4,7 @@ import (
 	"bytes"
 	"fmt"
 
-	ginclient "github.com/G-Node/gin-cli/gin-client"
+	ginclient "github.com/G-Node/gin-cli/ginclient"
 	"github.com/G-Node/gin-cli/util"
 	"github.com/spf13/cobra"
 )
