+++ gincmd/keyscmd.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 	"time"
 
-	ginclient "github.com/G-Node/gin-cli/gin-client"
+	ginclient "github.com/G-Node/gin-cli/ginclient"
 	"github.com/G-Node/gin-cli/util"
 	"github.com/spf13/cobra"
 )
