+++ gincmd/lscmd.go
@@ -6,7 +6,7 @@ import (
 	"sort"
 	"strings"
 
-	ginclient "github.com/G-Node/gin-cli/gin-client"
+	ginclient "github.com/G-Node/gin-cli/ginclient"
 	"github.com/G-Node/gin-cli/util"
 	"github.com/spf13/cobra"
 )
