+++ gincmd/unlockcmd.go
@@ -1,7 +1,7 @@
 package gincmd
 
 import (
-	ginclient "github.com/G-Node/gin-cli/gin-client"
+	ginclient "github.com/G-Node/gin-cli/ginclient"
 	"github.com/G-Node/gin-cli/util"
 	"github.com/spf13/cobra"
 )
